% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/RedditExtractoR.R
\name{ExtractAttributes}
\alias{ExtractAttributes}
\title{Extract data attributes}
\usage{
ExtractAttributes(URL, wait_time = 2)
}
\arguments{
\item{URL}{a string or a vector of strings with the URL address of pages of interest}

\item{wait_time}{wait time in seconds between page requests. 2 by default and it is also the minimum (API rate limit).}
}
\value{
A data frame with post / thread date (post_date), comment date (comm_date), number of comments within a post / thread (num_comments), subreddit (subreddit)
upvote proportion (upvote_prop), post /thread score (post_score), author of the post / thread (author), user corresponding to the comment (user),
comment score (comment_score), controversiality (controversiality), comment (comment), title (title), post / thread text (post_text), URL referenced (link)
domain of the references URL (domain)
}
\description{
Extract data attributes
}
\examples{
\dontrun{
example_attr = ExtractAttributes(URL="reddit.com/r/gifs/comments/39tzsy/whale_watching")
}
}

