% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_subreddits.R
\name{find_subreddits}
\alias{find_subreddits}
\title{Find subreddits by keywords}
\usage{
find_subreddits(keywords)
}
\arguments{
\item{keywords}{A string representing your search query}
}
\value{
A data frame with obtained reddits
}
\description{
Search for subreddits and their attributes based on a keyword
}
\examples{
\dontrun{
find_subreddits("cats")
}
}
