% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_thread_content.R
\name{get_thread_content}
\alias{get_thread_content}
\title{Get thread contents of Reddit URLs}
\usage{
get_thread_content(urls)
}
\arguments{
\item{urls}{A vector of strings pointing to a Reddit thread}
}
\value{
A list with 2 data frames "threads" and "comments"
}
\description{
This function takes a collection of URLs and returns a list with 2 data frames:
1. a data frame containing meta data describing each thread
2. a data frame with comments found in all threads
}
\details{
The URLs are being retained in both tables which would allow you to join them if needed
}
