% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedditExtractoR.R
\name{reddit_content}
\alias{reddit_content}
\title{Extract data attributes}
\usage{
reddit_content(URL, wait_time = 2)
}
\arguments{
\item{URL}{a string or a vector of strings with the URL address of pages of interest}

\item{wait_time}{wait time in seconds between page requests. 2 by default and it is also the minimum (API rate limit).}
}
\value{
A data frame with structure / position of the comment with respect to other comments (structure), ID (id), post / thread date (post_date), 
comment date (comm_date), number of comments within a post / thread (num_comments), subreddit (subreddit)
upvote proportion (upvote_prop), post /thread score (post_score), author of the post / thread (author), user corresponding to the comment (user),
comment score (comment_score), controversiality (controversiality), comment (comment), title (title), post / thread text (post_text), URL referenced (link)
domain of the references URL (domain)
}
\description{
Extract data attributes
}
\examples{
\dontrun{
example_attr = reddit_content(URL="reddit.com/r/gifs/comments/39tzsy/whale_watching")
}

}

