% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_content.R
\name{get_user_content}
\alias{get_user_content}
\title{Find data relating to a Reddit user}
\usage{
get_user_content(user)
}
\arguments{
\item{user}{A string representing a valid Reddit user name}
}
\value{
A list with "about" (list), "comments" (data frame) and "threads" (data frame)
}
\description{
Given a valid Reddit user name, obtain a list consisting of
general information about the user, their comments and threads
}
\examples{
\dontrun{
get_user_content("memes")
}
}
