\name{RedditExtractoR}
\alias{RedditExtractoR}
\alias{RedditExtractoR}
\docType{package}
\title{
Reddit Data Extraction Toolkit
}
\description{
Reddit is an online bulletin board and a social networking website where registered users can submit and discuss content. This package uses Reddit API to retrieve thread URLs, comments, subreddits and user information. For more information about the usage of this package, please see the following GitHub page: https://github.com/ivan-rivera/RedditExtractor
}
\details{
\tabular{ll}{
Package: \tab RedditExtractoR\cr
Type: \tab Package\cr
Version: \tab 3.0.0\cr
Date: \tab 2015-06-14\cr
License: \tab GPL-3 \cr
}
The package contains a collection of functions for extracting threads of interest and their corresponding comments, as well as functions for analysing the structure of these threads.
}
\author{
Ivan Rivera

Maintainer: Ivan Rivera <ivan.s.rivera@gmail.com>
}
\keyword{reddit}
\seealso{
\url{https://www.reddit.com}
}
